+++ tests/DbTest.php
@@ -1,0 +1,65 @@
+<?php
+
+/*
+ * Cerberus IRCBot
+ * Copyright (C) 2008 - 2016 Stefan Hüsges
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; either version 3 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+ * for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Cerberus;
+
+use Doctrine\DBAL\DriverManager;
+
+class DbTest extends \PHPUnit_Framework_TestCase
+{
+    protected static $config;
+    protected static $database;
+    protected $db;
+
+    public static function setUpBeforeClass()
+    {
+        date_default_timezone_set('Europe/Berlin');
+        self::$config = parse_ini_file(Cerberus::getPath() . '/config.ini', true);
+        self::$database = self::$config['testdb']['dbname'];
+    }
+
+    protected function setUp()
+    {
+        self::$config['testdb']['dbname'] = null;
+        $db = DriverManager::getConnection(self::$config['testdb']);
+        $sm = $db->getSchemaManager();
+        $sm->dropAndCreateDatabase(self::$database);
+        $db->close();
+        self::$config['testdb']['dbname'] = self::$database;
+        $this->db = DriverManager::getConnection(self::$config['testdb']);
+        $this->db->query(file_get_contents(Cerberus::getPath() . '/cerberus.mysql.sql'));
+        self::$config['db'] = self::$config['testdb'];
+    }
+
+    protected function tearDown()
+    {
+        if ($this->db === null) {
+            $this->fail('No connection to database...');
+        }
+        $sm = $this->db->getSchemaManager();
+        $sm->tryMethod('dropDatabase', self::$database);
+        $this->db->close();
+    }
+
+    public function testDb()
+    {
+
+    }
+}